Split profile search
